Got rid of WAClient, deprecated code. Prep for V3
Layered classes based on hierarchy as well.
