--- src/WAClient/WAClient.ts
@@ -1,7 +1,0 @@
-import WhatsAppWebMessages from './Messages'
-
-export { WhatsAppWebMessages as WAClient }
-export * from './Constants'
-export * from './Utils'
-export * from '../WAConnection/Constants'
-export { Browsers } from '../WAConnection/Utils'
